Made changes.

+++ src/main/java/org/infinite/spoty/viewModels/RequisitionDetailViewModel.java
@@ -208,7 +208,7 @@ public class RequisitionDetailViewModel {
 
             Platform.runLater(
                 () -> {
-                  requisitionDetailList.remove((int) getTempId());
+                  requisitionDetailList.remove(getTempId());
                   requisitionDetailList.add(getTempId(), requisitionDetail);
                 });
